move stories and docs file in folder
