+++ stories/Select/Guideline.mdx
@@ -7,7 +7,7 @@ import {
   Unstyled,
   Source,
 } from '@storybook/blocks';
-import { Select } from '../src/lib/components/selectv2/Selectv2.component';
+import { Select } from '../../src/lib/components/selectv2/Selectv2.component';
 
 import * as SelectStories from './selectv2.stories';
 
@@ -17,7 +17,7 @@ import * as SelectStories from './selectv2.stories';
 
 The Select component is a dropdown menu that allows users to choose from a list of options.
 It comes with various features to enhance user experience and adapt to different scenarios.
-You can use this [playground](http://localhost:3000/?path=/story/components-inputs-select--playground) to build the component
+You can use this [playground](?path=/story/components-inputs-select--playground) to build the component
 
 ## Size & Style
